# backend/assets/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "# Nexus Tasks REST API\n\nNexus Tasks is a collaborative project and task management platform.\nThis document describes the full REST API surface of the backend service.\n\n## Authentication\n\nMost endpoints require an authenticated session established via the `/api/v2/auth/login`\nor `/api/v2/auth/register` endpoints. Authentication state is maintained via an\nHTTP-only session cookie (`nexus_user`). Alternatively, machine clients may\nauthenticate using an API Key passed in the `X-API-Key` request header.\n\n## Base URL\n\nAll versioned resources are prefixed with `/api/v2`.",
    "title": "provides the Swagger/OpenAPI specification metadata for the Nexus Tasks API.",
    "version": "2.0.0"
  },
  "host": "localhost:8000",
  "basePath": "/api/v2",
  "paths": {
...
    "/inbox": {
      "get": {
        "description": "Returns all unread and uncleared notifications for the authenticated user, ordered by most recent first.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "notifications"
        ],
        "summary": "Get notification inbox",
        "operationId": "getInbox",
        "responses": {
          "200": {
            "description": "List of notifications."
          },
          "401": {
            "description": "Not authenticated."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/inbox/clear-all": {
      "patch": {
        "description": "Marks all of the authenticated user's notifications as cleared in a single operation.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "notifications"
        ],
        "summary": "Clear all notifications",
        "operationId": "clearAllNotifications",
        "responses": {
          "200": {
            "description": "All notifications cleared."
          },
          "401": {
            "description": "Not authenticated."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/inbox/{notificationId}/clear": {
      "patch": {
        "description": "Marks a notification as cleared, removing it from the user's active inbox view.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "notifications"
        ],
        "summary": "Clear a notification",
        "operationId": "clearNotification",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "description": "UUID of the notification.",
            "name": "notificationId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Notification cleared."
          },
          "400": {
            "description": "Invalid notification ID."
          },
          "401": {
            "description": "Not authenticated."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/inbox/{notificationId}/read": {
      "patch": {
        "description": "Sets the read flag on the specified notification. The notification remains visible until cleared.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "notifications"
        ],
        "summary": "Mark a notification as read",
        "operationId": "markNotificationRead",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "description": "UUID of the notification.",
            "name": "notificationId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Notification marked as read."
          },
          "400": {
            "description": "Invalid notification ID."
          },
          "401": {
            "description": "Not authenticated."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
...
  },
...
  "securityDefinitions": {
    "apiKeyAuth": {
      "description": "A long-lived API key generated via the `/api/v2/auth/api-keys` endpoint. Intended for machine-to-machine access.",
      "type": "apiKey",
      "name": "X-API-Key",
      "in": "header"
    },
    "cookieAuth": {
      "description": "HTTP-only session cookie set by the login/register endpoints. The cookie name is `nexus_user` and contains a signed JWT.\n",
      "type": "apiKey",
      "name": "Cookie",
      "in": "header"
    }
  },
  "security": [
    {
      "cookieAuth": []
    },
    {
      "apiKeyAuth": []
    }
  ]
}